requirements.txt
coverage==6.5.0
flask==2.2.2
Pillow==9.2.0
pycodestyle==2.9.1
pydocstyle==6.1.1
pyre-check==v0.9.11
pytest==7.1.3
python-docx==0.8.11
requests==2.28.1